component Ratings moved to tiles folder
